feat(BallNumber): 공의 숫자를 검증하는 로직 구현
