+++ src/test/java/baseball/domain/BallTest.java
@@ -22,15 +22,6 @@ class BallTest {
     }
 
     @ParameterizedTest
-    @ValueSource(ints = {0, 10})
-    @DisplayName("공의 숫자가 1부터 9까지의 범위를 벗어나면 예외가 발생한다.")
-    void invalidNumber(int ballNumber) {
-        assertThatThrownBy(() -> new Ball(ballNumber, 1))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessageContaining("공의 숫자는 1에서 9사이의 값이어야 합니다.");
-    }
-
-    @ParameterizedTest
     @ValueSource(ints = {0, 4})
     @DisplayName("공의 위치가 1부터 3까지의 범위를 벗어나면 예외가 발생한다.")
     void invalidPosition(int position) {
